chore(tooling): phase 2b — split content script, drop extension-package (issue #56)
ADR 0003: declare src/ canonical and delete extension-package/. The
two trees had drifted by ~678 lines (all formatting + JSDoc Object→object
fixes — no behavioral content unique to extension-package/). scripts/build.js
now reads from src/ instead.

ADR 0004: split contentScript.js (3399 lines) into four Manifest V3 content
scripts sharing one isolated world:
  - contentScript.js (1399 lines) — state, overlay, scan orchestration,
    keyboard navigation
  - elementChecks.js (410 lines) — per-element accessibility checks
  - uiPanels.js (1152 lines) — filter/summary/config/progress panels,
    categorizeIssue, analyzeLogs
  - reportGenerators.js (548 lines) — JSON/CSV/HTML/text export

manifest.json lists all four in order. Cross-file refs use /* global */
ESLint directives (with :writable where reassigned). Jest test contract
preserved: contentScript.js requires the siblings in test mode and exposes
state on \`global\` first so the CommonJS-isolated sibling modules can see
it. All four files share the same NODE_ENV=test global-export idiom.

eslint.config.js: add max-lines (1500), max-lines-per-function (250),
complexity (40), max-depth (6), max-nested-callbacks (5), max-params (5)
project-wide. Test files override size/nesting gates since Jest describe()
callbacks legitimately wrap entire files. Also allowlist Firefox WebExt
manifest keys (browser_specific_settings, strict_min_version) for the
camelcase rule.

263 unit tests pass; Chrome/Firefox/Edge builds produce valid zips.
