+++ .prettierignore
@@ -2,7 +2,6 @@ node_modules
 dist
 build
 coverage
-extension-package
 scratchpads
 reports
 *.min.js
