--- extension-package/manifest.json
@@ -1,48 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Accessibility Highlighter",
-  "version": "1.0.3",
-  "description": "A browser extension to highlight accessibility problems on the page, giving a visual indication of what the problems are",
-  "author": "AFixt",
-  "permissions": ["storage", "activeTab"],
-  "host_permissions": ["*://*/*"],
-  "background": {
-    "service_worker": "background.js"
-  },
-  "action": {
-    "default_icon": {
-      "16": "icons/icon-16.png",
-      "48": "icons/icon-48.png",
-      "128": "icons/icon-128.png"
-    },
-    "default_title": "Accessibility Highlighter - Click to toggle accessibility checking"
-  },
-  "content_scripts": [
-    {
-      "matches": ["<all_urls>"],
-      "js": ["contentScript.js"],
-      "run_at": "document_idle"
-    }
-  ],
-  "icons": {
-    "16": "icons/icon-16.png",
-    "48": "icons/icon-48.png",
-    "128": "icons/icon-128.png"
-  },
-  "web_accessible_resources": [{
-    "resources": ["icons/*"],
-    "matches": ["<all_urls>"]
-  }],
-  "content_security_policy": {
-    "extension_pages": "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data:; font-src 'self'; connect-src 'none'; object-src 'none'; media-src 'none'; frame-src 'none'; base-uri 'none'; form-action 'none'; frame-ancestors 'none'; upgrade-insecure-requests;"
-  },
-  "commands": {
-    "toggle-accessibility": {
-      "suggested_key": {
-        "default": "Ctrl+Shift+A",
-        "mac": "Command+Shift+A"
-      },
-      "description": "Toggle accessibility highlighting on/off"
-    }
-  }
-}
