+++ manifest.json
@@ -20,7 +20,7 @@
   "content_scripts": [
     {
       "matches": ["<all_urls>"],
-      "js": ["contentScript.js"],
+      "js": ["contentScript.js", "elementChecks.js", "uiPanels.js", "reportGenerators.js"],
       "run_at": "document_idle"
     }
   ],
